Fix: オブジェクトが削除できない問題を修正

+++ renderer/js/map/point/point.js
@@ -14,18 +14,20 @@ export function initPoints() {
     return hash;
   }
 
-  points.delete = hash => {
-    for (let item of points[hash].linkedObjects) {
-      let obj = objects[item];
-      obj.linkedPoints.splice(obj.linkedPoints.indexOf(hash), 1);
+  points.delete = (hash, removeFromObject) => {
+    if (removeFromObject) {
+      for (let item of points[hash].linkedObjects) {
+        let obj = objects[item];
+        obj.linkedPoints.splice(obj.linkedPoints.indexOf(hash), 1);
 
-      //中継点の数がポイントの場合0個、ラインの場合1個以下、エリアの場合2個以下のときオブジェクトを削除
-      if (obj.form == 'point' && obj.linkedPoints.length == 0) objects.delete(item);
-      else if (obj.form == 'line' && obj.linkedPoints.length <= 1) objects.delete(item);
-      else if (obj.form == 'area' && obj.linkedPoints.length <= 2) objects.delete(item);
-      else {
-        objects.update(item);
-        drawObject(obj);
+        //中継点の数がポイントの場合0個、ラインの場合1個以下、エリアの場合2個以下のときオブジェクトを削除
+        if (obj.form == 'point' && obj.linkedPoints.length == 0) objects.delete(item);
+        else if (obj.form == 'line' && obj.linkedPoints.length <= 1) objects.delete(item);
+        else if (obj.form == 'area' && obj.linkedPoints.length <= 2) objects.delete(item);
+        else {
+          objects.update(item);
+          drawObject(obj);
+        }
       }
     }
